facil locator actions restructured
