+++ src/applications/facility-locator/actions/clearGeocodeError.js
@@ -1,0 +1,5 @@
+import { GEOCODE_CLEAR_ERROR } from '../utils/actionTypes';
+
+export const clearGeocodeError = () => async dispatch => {
+  dispatch({ type: GEOCODE_CLEAR_ERROR });
+};
