+++ src/applications/facility-locator/actions/clearSearchResults.js
@@ -1,0 +1,5 @@
+import { CLEAR_SEARCH_RESULTS } from '../utils/actionTypes';
+
+export const clearSearchResults = () => ({
+  type: CLEAR_SEARCH_RESULTS,
+});
