+++ src/applications/facility-locator/actions/clearSearchText.js
@@ -1,0 +1,5 @@
+import { CLEAR_SEARCH_TEXT } from '../utils/actionTypes';
+
+export const clearSearchText = () => async dispatch => {
+  dispatch({ type: CLEAR_SEARCH_TEXT });
+};
