+++ src/applications/facility-locator/actions/mapMoved.js
@@ -1,0 +1,6 @@
+import { MAP_MOVED } from '../utils/actionTypes';
+
+export const mapMoved = currentRadius => ({
+  type: MAP_MOVED,
+  currentRadius,
+});
